startupAnimation.js: Make sure the shroud actor remains the size of the stage during startup.
As a virtual machine guest, the 'monitor' may be resized after the
shroud actor is set up, but before (or during) the animation itself,
causing the shroud to only cover a portion of the screen.

The constraints will make sure the shroud resizes if the stage does
at any point during startup.