fix: refactor: separate icon theme dir creation
Like it was previously with all the both directories being used in the same try/except block, whenever the first call to os.makedirs failed (this happens if the directory was already created when trying to create it), the second directory would never be created. The solution was to create each icon dir separately.