panel-launchers: Don't allow the size of custom icons to be reset.
When the theme changed, _updateIconSize would compare the current
icon size with smallest of either its container's height or its
width.

Only one direction would be set (depending on panel orientation),
so the smallest size was always 0, resetting the StIcon's icon_size
property to 0. This would force it to look at the theme for the
icon size (which would be too large for the panel).