cinnamon-desktop-editor: Use a path when setting a panel-launcher icon.
If an icon name is used in the panel launcher applet, and the user
switches to an icon theme that doesn't have that icon name, the
launcher will disappear.

Now, the selected icon's path is checked - if it's coming from the
hicolor theme, it's left as-is (since hicolor is inherited by the
active theme). Otherwise, the icon's path is stored instead of its
name, so it will remain valid regardless of future themes.

Add an optional 'icon-size' argument to cinnamon-desktop-editor,
so the path to an appropriately-sized icon is used for the launcher.