Add keyboard shortcuts for warping the mouse pointer between monitors.
When working with multiple monitors, it is tedious and fatiguing
to move your pointer manually when changing focus from one monitor
to another - especially if you have more than two.

This will enable the user to assign shortcuts to warp the pointer
to the next and previous logical monitors.

An additional setting will specify either centering the pointer
on the new monitor, or maintaining the relative position it had on
the previous one.