Add class to make applets resizable with mouse
When applet can be resized on opposite edges, i.e. when the applet is not in a corner of the workspace, then resizing from either of those edges leads the mouse pointer to not match up with the applet edge. This is a minor visual issue that maybe should be fixed in the future.