cinnamon-settings: Make sure the theme correctly falls back to hicolor when looking up icons.
When using 'icon-name' on the pixbuf renderer, Gtk attempts to
shorten hyphenated names in order to find a match in the current
them *before* checking the hicolor theme for an exact match.

For example, 'cs-gestures' was missing from the active theme.
Instead of looking for 'cs-gestures' in the hicolor theme (which
is why the hicolor theme exists), it looks for an icon named 'cs',
which happens to be an actual icon name in the current theme.

Populate the store with GIcons instead. GThemedIcon performs the
fallback lookups in a more sensible order, and gives us the correct
hicolor icon.

Fixes https://github.com/linuxmint/mint21.2-beta/issues/8