theme: Don't force too large a minimum size of PopupSliderMenuItem.
This should be smaller, or removed entirely, but just making it
smaller will prevent too many side-effects on xlets relying on a
minimum width.

This keeps narrower popup menus from being forced larger when a
submenu is opened, which can be jarring. As an example, this occurs
in the right-click->Applications expander in the sound applet.

We use sliders in the power applet also, bump that up a bit from our
new minimum, it looks nicer.

Ref: linuxmint/mint21.2-beta#36