Fix thumbnails not appearing for windows created while menu is open (#11759)
With this change, I don't have #11106 happen anymore.

For me, the bug happened whenever a window is created while the hover
menu is open.

Thumbnails objects were being created and added to the list of
thumbnails in the AppThumbnailHoverMenu. However, then refreshThumbnail
is called, which calls getThumbnail. In getThumbnail, metaWindowActor
is still null and metaWindow.get_compositor_private returned null,
causing it to remove the thumbnail again before it was ever visible.

With the setTimeout(..., 0), getThumbnail is called on the next
event loop turn, so that the actor will be returned by
get_compositor_private.

If the menu isn't open when the window is created, then the window is
added to the queuedWindows array and added later on when the actor is
available.