keybindings: Make Above_Tab displayable.
The "Above_Tab" descriptor only exists in muffin's keybinding
handling. Keyboard layouts are all different but the position
should be the same for the default switch-group and switch-group-
backward keybindings. When muffin handles an action using this
Above_Tab assignment, it will figure out what that key is
for the current keyboard layout automatically.

If you re-teach those keybindings using the key above your
tab, it will be displayed in the proper form (Alt+` for US), and
the automation above will no longer occur - the keybinding will be
layout specific. Resetting it to default (right-click the keybinding
description) will restore the original behavior.

This change is only visual to show that the keybindings aren't
actually 'unassigned' (because gtk_accelerator_parse_with_keycode()
doesn't know what Above_Tab means).

Fixes #11778.