window-list/gwl: Simplify window focus updates.
Instead of starting at the primary window and iterating thru
transients to check for focus, check the MetaDisplay's focus-window
to see if the primary window is an ancestor.

Cjs seems to have an issue nesting the meta_window_foreach_transient
closure inside the notify closure for the window 'appears-focused'
property. This causes a reference count warning for GClosure, and
occasionally a crash.

This is reproducible with a normal GtkDialog with a transient
parent set, and toggle the window list item for the application.