Adds option to always show the input switch in the sound applet. (Closes #11232)
The PR https://github.com/linuxmint/cinnamon/pull/10620/ hides the input
switch, if not recording device is open. But let's say you want to mute
your microphone before starting a video session, you need that input
switch _before_ you start the conference software!

This adds a switch to the configuration screen to always enable the
input switch. The default is the old behaviour: Only show it, when a
recording device is open. (But you may want to change that.)

This is based on the
https://github.com/linuxmint/cinnamon-spices-applets/commit/670c97dc987c42b8c13da9d8cb034bd0f080577a
for the sound150@claudiux spice.