cs-info: increase inxi verbosity (#11984)
Add a third 'x' option to the inxi command so it will also provide the
partitioning scheme for drives. It is useful for people helping with
installation or boot issues to know upfront whether GPT or MBR is used.

Also adds the 'J' option to make inxi provide information about USB
devices, to put it on par with the information mintreport provides.