cs_keyboard: Enable configuration of bindings for Spices
This adds the Spices category to the Keyboard module and prepends a
category label in binding search results. When Spices are selected (or
instances, where applicable), the applicable Spice will be highlighted
(where possible) to indicate which Spice the binding applies to. This
also changes the functionality such that the results from a binding
search jump the the category for the binding once it is selected.