cs_themes: Blacklist Adwaita and HighContrast
Adwaita looks awful since GNOME 46. It removed category icons
and a huge number of icons used by Cinnamon. This theme is
only meant to be used in GNOME and does not support Cinnamon.

HighContrast is very incomplete also. It's also a global setting
in A11Y.

GTK depends on Adwaita so these themes are always present. This commit
removes them from the theme selection since they make Cinnamon look
broken.

Note: This was done in Budgie in 2022 for the same reasons. It should
probably be done in all DEs really.
https://github.com/BuddiesOfBudgie/budgie-desktop/commit/27b944769ada26e4d8c60252a4602f7b15b2688f