cs_themes: Blacklist a few more themes
Cleaning up on a few additional themes:

- Papirus looks really good but it ships with ePapirus which you can't
remove on its own. ePapirus is designed specifically for Pantheon, so
there's no reason to choose it over Papirus in Cinnamon.

- ubuntu-mono/humanity. Ubuntu made GTK depend on them so they're always
in the list, looking out of place in derivatives and impossible to
remove for end-users. They're there for inheritance purpose only, the
Ubuntu desktop (GNOME-based) won't let you choose them and defaults
to a much better looking theme: Yaru.