userWidget.js: Fix AccountsService icon changes.
In many instances, calling user.set_icon_file() will copy the
actual file into /var/lib/AccountsService/icons.

The name of this file is always the same as the username. The
problem with this is that Cinnamon's texture cache uses lookup keys
based partly on the serialization of a GIcon.

For GFileIcons, the resulting string is based on the file's path.
Since the path doesn't change in these cases, Cinnamon will continue
to use the old icon.

- Change how the avatar image is loaded, so that it doesn't get
  cached by Cinnamon.
- Try to use the theme icon when possible (this allows for better
  symbolic icon support).
- Add a method to resize the avatar image, so this can be used
  in panels, where the allowed size is subject to change.