userWidget: Back to using the theme to load the face image.
This only works properly when the image is set as a background
as part of the theme, as some themes want to frame the image.

Go back to using the original file picker for an image - simple
icons don't really look very nice as avatar pics, and that was
the only reason to involve the xapp icon picker.