xdg-desktop-portal: Don't ship the x-cinnamon-portals.conf file.
This is the 'new' way of declaring portal support, but the order
of precedence that this was supposed to fix is currently broken,
and the Gtk portal services some requests that our implementation
(xdg-desktop-portal-xapp) should be handling.

The xdg-destop-portal-xapp package still installs the 'deprecated'
xapp.portal file, with which everything still works as it should.

We can re-add this file once we confirm upstream has fixed the
ordering issue.

Ref:
https://github.com/flatpak/xdg-desktop-portal/pull/1240