cs_keyboard.py: Skip invalid bindings (#12396)
It is possible to remove Spices manually by removing the associated directories
rather than using the appropriate module. This results in the bindings still
being stored internally. When loading the Keyboard module and this scenario has
occurred, it results in a failure due to the matching files having been
removed. We leave the mappings internally so that if the matching Spice is
reinstalled, the previous settings will be retained.