modalDialog: Destroy on most dialogs on close
We don't need to keep these hanging around when we are done with them. In
addition, most dialogs get their initial focus set during construction. By
recreating the dialogs on each use we ensure we get the proper focus set.
One exception is the run dialog so we have a parameter to avoid destroying
dialogs we don't want to.