cs_display.py: Expose option to choose fractional scaling direction.
This is a feature that was broken since muffin 5.4, but has since
been fixed in https://github.com/linuxmint/muffin/pull/694.

This can be useful in multi-monitor situations where there is a
mix of normal and hidpi monitors.

Scenario:
- monitor 0:  1920x1080
- monitor 1:  3840x2160

In the default scale-down mode, in order to achieve a desirable
configuration, the global UI scale must be doubled (2x), then
monitor 0 must be scaled back down to 100%. This results in
deteriorated performance and visuals in monitor 0.

In scale-up mode, the global scale can remain at 1x, and monitor 1
is 'fractionally' scaled to 200%. Since the scale factor is exactly
double, there is no loss in quality for monitor 1 and performance
impact is minimal.