polkitAuthenticationAgent.js: Retry pushModal if necessary.
Double-clicking a folder that requires authentication to open
can become a race condition that causes pushModal to fail if
the 'grab' generated by the application isn't released quickly
enough. (Hitting 'enter' is a workaround here).

This is similar to a situation that can be encountered when
when activating the screensaver using keyboard shortcuts. That
uses a similar workaround.