polkitAuthenticationAgent.js: Handle system accounts properly. (#12865)
* polkitAuthenticationAgent.js: Handle system accounts properly.

Following up 9c8a15b62f573, user accounts marked as system accounts
in AccountsService needed to be treated similarly to root accounts
when authenticating in a non-admin session.

Fixes linuxmint/linuxmint#765.

* polkitAuthenticationAgent.js: Retry pushModal if necessary.

Double-clicking a folder that requires authentication to open
can become a race condition that causes pushModal to fail if
the 'grab' generated by the application isn't released quickly
enough. (Hitting 'enter' is a workaround here).

This is similar to a situation that can be encountered when
when activating the screensaver using keyboard shortcuts. That
uses a similar workaround.