Calender: Fix issue with with custom tooltip not updating seconds. (#12786)
When the calender applet uses a custom format, and the tooltip contains
seconds but the regular custom format doesn't,  the tooltip seconds
wouldn't update properly.

This is because the update frequency is based on the regular custom
format.

This has been fixed, so that the calender now checks whether the mouse
has entered the applet. When the tooltip is shown, the format string
used by the clock is updated to combine the tooltip format. This way it
ensures that the update frequency is based on whichever has the finer
resolution.

Fixes #12140