Don't hide cinnamon's chrome if desktop window is focused
This commit ensures that chrome isn't hidden if the desktop window is the last focused window on the monitor. This is useful when chrome needs to be shown over a fullscreened window and remain visible until the fullscreened window is again focused by the user. This is done by focusing the desktop window, e.g. when opening an applet with a shortcut key over a fullscreened window.