determine the shutdown dialog default action
This reads the string array from GSesttings.

While keeping the order, it filters out any values which are either completely
unknown or whose action is not available (as determined by the `can*`-
variables; the `cancel`-action is always available). Duplicate values don’t
matter and are thus left as is.

If the resulting array is empty, subscribing its first element will yield
`undefined`, which is just as desired and causes no default action to be set at
all.

Signed-off-by: Christoph Anton Mitterer <mail@christoph.anton.mitterer.name>