GWL: Don't show places
These are available in one click from:

- menu@cinnamon.org
- removable-drives@cinnamon.org
- nemo's context menu (actions) in GWL
- nemo's sidebar

It makes no sense to have them there in one more click, for nemo
specifically, burried under a subcategory and configurable with
"show recent" which shouldn't be related to it.

And then if we support recents, and places, why not do favorites?
This does not belong here.