menu: Toggle between symbolics or fullcolor for categories
This used to be an all or nothing preference. Turning it off allowed
more focus on the apps but it made the category box tiny and not very
good looking.

This commit replaces this with subtle symbolic icons. They're small,
they're all the same (same icon as for all-applications). The section
looks good but doesn't draw the user's attention. This is the new
default.