menu: Make places/bookmarks configurable
This is important because:

- distributions want to be able to fill the menu with a number of
favs and places which fit nicely in the default menu size.

- Users can have a lot of bookmarks and might not want to see them
in the menu, but special dirs are bookmarks, so it's all or nothing..

- Some special dirs are more important than others, based on preference,
for instance you might never use Music, but use Downloads all the time,
other people don't..

- Just like distributions, users will want to fill their sidebar with
places and favs with the right mix and the right number of items which'
fits nicely. Having the ability to decide what is shown and what isn't
lets them do just that.