menu: Use new CSS names
Use different class names than the old menu.

That way we can update themes with the new style section without
having to affect the section for the old menu. This means the theme
can be updated for both new and old cinnamon versions, without having
to get versioned.

If people fork the old menu and bring it to spices, themes will
continue to work with it without needing it to get new class names
and support.