avatar: Apply a clutter effect on hover
This allows us to continue to make the avatars into a circle but
without applying a background and a border.

It looks much neater as a result and takes full advantage of the
avatar when its picture has a transparent background.

This also simplifies things for the themes and the applets. The icon
reacts both in the menu and the user applets, or anywhere else
it is used.