Switch to XApp symbolic icons
Note: Because our JS layer allows us to refer to symbolic icons
using fullcolor names, we don't have exact matches in JS and it's hard
to find all the context matches.

This commit probably tackles all the C and Python layers, but we'll need
ongoing efforts to migrate all the icons from the JS layer.

It's not urgent and the work is done in XApp so this can happen slowly.