ws: Repaint the workspace state during the minimize
If the window is minimized, we need to trigger a repaint of the workspace graph here because there's an edge case where if this window is minimized out of focus (i.e. while it isn't the focused window), the graph won't update properly update to reflect the state of other underlying windows.