workspace switcher: Add option to display window icons (#12202)
* ws: Add option to display window icons

* ws: Made some small refactoring on the codes

* ws: Hide the icon if there isn't enough space for it

* ws: turn icon size selection into a combobox and allow auto adjust of the size

* ws: optimize intersection rect calcs

* ws: don't set current focus window by default during instantiation

This is handled after during button creation, and can lead to workspace switcher applet not following the focus windows initially.

* ws: improve icon handling

* ws: Update all workspaces for windows in all workspaces

* ws: Use strict comparison in workspace graph update

* ws: Refactor WorkspaceGraph repaint to improve memory management

* ws: Resolve suggestions

* ws: Show icon and auto-size by default

Remove window icon settings and refactor icon handling in WindowGraph

* ws: Only render unobscured windows in the workspace graph

ref: https://github.com/linuxmint/muffin/pull/766

* ws: Improve window state representation and management

* ws: Avoid triggering a whole refresh on mouse hover

* ws: Repaint the workspace state during the minimize

If the window is minimized, we need to trigger a repaint of the workspace graph here because there's an edge case where if this window is minimized out of focus (i.e. while it isn't the focused window), the graph won't update properly update to reflect the state of other underlying windows.