magnifier.js: Revert part of 894ab13f - go back to using Cinnamon's KeybindingManager to setup zoom keybindings.
Previously zoom shortcuts were usable in any state - modal or non-
modal, because Main._stageEventHandler matched up action ids with
the keybinding manager's ids to invoke the correct callbacks.

When we switched to global.display (in order to gain automatic
updating when the keybinding changed) we lost that capability.

Go back to that, but make sure to update these magnifier keybindings
when their settings change.