main.js: Don't try to invoke keybindings for synthetic events.
The 'filter_key_event' in inputMethod.js can now generate duplicate
events from editable ClutterText actors. This was causing our
global keybindings to fire twice when a modal popup with an entry
was focused.

To reproduce: Open the Cinnamon menu, or Alt-F2, and try to take
a screenshot - you end up with two.