keyboard: Always fetch the full XkbInfo store except when adding a layout.
This adapts Cinnamon's use of CinnamonDesktop.XkbInfo to
linuxmint/cinnamon-desktop@2bb17da29eb, which stops us relying
on a setting to load a full XkbInfo model.

This setting will still be used directly in the Add Layout dialog,
but it would cause problems if the user added an 'extra' layout,
then later un-set the 'show-all-sources' toggle. Cinnamon would
consider it an invalid layout.