Put back "notifications" role removed in commit...
... https://github.com/linuxmint/cinnamon/commit/1c995f606972dddd339d6dfbc13dc908bdf757cb

While the "notifications" role is no longer used in cinnamon 6.6, removing the role in cinnamon prevents 3rd party applets from loading if the role is listed in the 3rd party's metadata.json file. While the 3rd party applet can be versioned to adapt to the changes in cinnamon, the metadata.json file cannot. Therefore the applet should not be prevented from loading even though "notifications" role is no longer used in cinnamon.