st-entry.c: Disable cursor blink.
There is an issue with ClutterText not receiving the correct
text direction, and creating an invalid layout, which results in
a crash if the user types in a RTL language while the cursor
is hidden.

Muffin currently relies on gtk3 translations to get the correct
direction. These may not even be installed if the user is only
using a RTL keyboard layout, and not the actual language on their
system. Also, gtk3 translations rely on language-pack-gnome* in
Mint/Ubuntu:

const char *e = g_dgettext ("gtk30", "default:LTR");

For now, disable blinking entirely for StEntry actors until this
source of confusion can be eliminated.

ref: #13086, #12248, #12949, #11220, #12816, #12753.