keyboardManager.js: Make sure to reload the current engine properties when sources change.
If an ibus source is first, when a reload occurs due to sources
changing, Cinnamon clears out existing InputSources and any
ibus ui properties. But if that source was already set as the
global engine, those properties won't be automatically reloaded.

Trigger a reload during activation even if the current source hasn't
changed.